requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pyodbc==5.1.0
